requirements.txt
fastapi==0.112.0
uvicorn==0.30.0
pydantic==2.7.4
starlette==0.37.2
smolagents
openai>=1.30.0
python-docx